moved app to components because it was dumb
